backend/data-analysis/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.8.6
pydantic==2.5.3
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1
numpy==1.24.3
scipy==1.11.4
matplotlib==3.7.2
pandas==2.0.3